+++ docs/tr1/CHANGELOG.md
@@ -44,6 +44,7 @@
 - fixed header and arrows disappearing when the inventory ring rotates (#2352, regression from 4.4)
 - fixed Story So Far feature not playing opening FMVs from the current level (#2360, regression from 4.2)
 - fixed `/play` command crashing the game when used after loading a level (#2411, regression)
+- fixed various death counter problems (existing saves will have the count reset) (#2412, regression from 2.6)
 - fixed `/demo` command crashing the game if no demos are present (regression from 4.1)
 - fixed Lara not being able to jump or stop swimming if the related responsive config options are enabled, but enhanced animations are not present (#2397, regression from 4.6)
 - improved pause screen compatibility with PS1 (#2248)